- fixed requested changes
