+++ x/posts/internal/keeper/common_test.go
@@ -46,11 +46,9 @@ func testCodec() *codec.Codec {
 var testPostOwner, _ = sdk.AccAddressFromBech32("cosmos1y54exmx84cqtasvjnskf9f63djuuj68p7hqf47")
 var timeZone, _ = time.LoadLocation("UTC")
 var testPostCreationDate = time.Date(2020, 1, 1, 15, 15, 00, 000, timeZone)
-var medias = types.PostMedias{
-	types.PostMedia{
-		URI:      "https://uri.com",
-		MimeType: "text/plain",
-	},
+var medias = types.PostMedia{
+	URI:      "https://uri.com",
+	MimeType: "text/plain",
 }
 
 var testPost = types.NewPost(
@@ -62,5 +60,5 @@ var testPost = types.NewPost(
 	map[string]string{},
 	testPostCreationDate,
 	testPostOwner,
-	medias,
+	types.PostMedias{medias},
 )
