+++ x/posts/internal/keeper/handler_test.go
@@ -33,7 +33,7 @@ func Test_handleMsgCreatePost(t *testing.T) {
 					testPost.AllowsComments,
 					"desmos",
 					map[string]string{},
-					testPost.LastEdited,
+					testPost.Created,
 					testPost.Creator,
 					testPost.Medias,
 				),
@@ -46,7 +46,7 @@ func Test_handleMsgCreatePost(t *testing.T) {
 				"desmos",
 				map[string]string{},
 				testPost.Creator,
-				testPost.LastEdited,
+				testPost.Created,
 				testPost.Medias,
 			),
 			expError: "Post with id 1 already exists",
@@ -60,7 +60,7 @@ func Test_handleMsgCreatePost(t *testing.T) {
 				testPost.Subspace,
 				testPost.OptionalData,
 				testPost.Creator,
-				testPost.LastEdited,
+				testPost.Created,
 				testPost.Medias,
 			),
 			expPost: types.NewPost(
@@ -70,7 +70,7 @@ func Test_handleMsgCreatePost(t *testing.T) {
 				testPost.AllowsComments,
 				testPost.Subspace,
 				testPost.OptionalData,
-				testPost.LastEdited,
+				testPost.Created,
 				testPost.Creator,
 				testPost.Medias,
 			),
@@ -84,7 +84,7 @@ func Test_handleMsgCreatePost(t *testing.T) {
 				"desmos",
 				map[string]string{},
 				testPost.Creator,
-				testPost.LastEdited,
+				testPost.Created,
 				testPost.Medias,
 			),
 			expError: "Parent post with id 50 not found",
@@ -99,7 +99,7 @@ func Test_handleMsgCreatePost(t *testing.T) {
 					false,
 					"desmos",
 					map[string]string{},
-					testPost.LastEdited,
+					testPost.Created,
 					testPost.Creator,
 					testPost.Medias,
 				),
@@ -111,7 +111,7 @@ func Test_handleMsgCreatePost(t *testing.T) {
 				"desmos",
 				map[string]string{},
 				testPost.Creator,
-				testPost.LastEdited,
+				testPost.Created,
 				testPost.Medias,
 			),
 			expError: "Post with id 50 does not allow comments",
@@ -182,19 +182,19 @@ func Test_handleMsgEditPost(t *testing.T) {
 		{
 			name:       "Post not found",
 			storedPost: nil,
-			msg:        types.NewMsgEditPost(types.PostID(0), "Edited message", testPostOwner, testPost.LastEdited),
+			msg:        types.NewMsgEditPost(types.PostID(0), "Edited message", testPostOwner, testPost.Created),
 			expError:   "Post with id 0 not found",
 		},
 		{
 			name:       "Invalid editor",
 			storedPost: &testPost,
-			msg:        types.NewMsgEditPost(testPost.PostID, "Edited message", editor, testPost.LastEdited),
+			msg:        types.NewMsgEditPost(testPost.PostID, "Edited message", editor, testPost.Created),
 			expError:   "Incorrect owner",
 		},
 		{
 			name:       "Edit date before creation date",
 			storedPost: &testPost,
-			msg:        types.NewMsgEditPost(testPost.PostID, "Edited message", testPost.Creator, testPost.LastEdited.AddDate(0, 0, -1)),
+			msg:        types.NewMsgEditPost(testPost.PostID, "Edited message", testPost.Creator, testPost.Created.AddDate(0, 0, -1)),
 			expError:   "Edit date cannot be before creation date",
 		},
 		{
@@ -211,7 +211,7 @@ func Test_handleMsgEditPost(t *testing.T) {
 				Subspace:       testPost.Subspace,
 				OptionalData:   testPost.OptionalData,
 				Creator:        testPost.Creator,
-				Medias:         medias,
+				Medias:         testPost.Medias,
 			},
 		},
 	}
