+++ x/posts/internal/types/codec.go
@@ -17,5 +17,4 @@ func RegisterCodec(cdc *codec.Codec) {
 	cdc.RegisterConcrete(MsgEditPost{}, "desmos/MsgEditPost", nil)
 	cdc.RegisterConcrete(MsgAddPostReaction{}, "desmos/MsgAddPostReaction", nil)
 	cdc.RegisterConcrete(MsgRemovePostReaction{}, "desmos/MsgRemovePostReaction", nil)
-	cdc.RegisterConcrete(Post{}, "desmos/Post", nil)
 }
