+++ x/posts/internal/types/post_response_test.go
@@ -36,6 +36,18 @@ func TestPostQueryResponse_MarshalJSON(t *testing.T) {
 		postOwner,
 		medias,
 	)
+	postNoMedia := types.NewPost(
+		types.PostID(10),
+		types.PostID(0),
+		"Post",
+		true,
+		"desmos",
+		map[string]string{},
+		date,
+		postOwner,
+		types.PostMedias{},
+	)
+
 	likes := types.Reactions{
 		types.NewReaction("like", liker),
 		types.NewReaction("like", otherLiker),
@@ -50,9 +62,14 @@ func TestPostQueryResponse_MarshalJSON(t *testing.T) {
 		expResponse string
 	}{
 		{
-			name:        "Post Query Response with Post",
+			name:        "Post Query Response with Post that contains media",
 			response:    PostResponse,
-			expResponse: `{"post":{"id":"10","parent_id":"0","message":"Post","created":"2020-02-02T15:00:00Z","last_edited":"0001-01-01T00:00:00Z","allows_comments":true,"subspace":"desmos","creator":"cosmos1y54exmx84cqtasvjnskf9f63djuuj68p7hqf47","medias":[{"uri":"https://uri.com","mime_Type":"text/plain"}]},"reactions":[{"owner":"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4","value":"like"},{"owner":"cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae","value":"like"}],"children":["98","100"]}`,
+			expResponse: `{"id":"10","parent_id":"0","message":"Post","created":"2020-02-02T15:00:00Z","last_edited":"0001-01-01T00:00:00Z","allows_comments":true,"subspace":"desmos","creator":"cosmos1y54exmx84cqtasvjnskf9f63djuuj68p7hqf47","medias":[{"uri":"https://uri.com","mime_Type":"text/plain"}],"reactions":[{"owner":"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4","value":"like"},{"owner":"cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae","value":"like"}],"children":["98","100"]}`,
+		},
+		{
+			name:        "Post Query Response with Post that not contains media",
+			response:    types.NewPostResponse(postNoMedia, likes, children),
+			expResponse: `{"id":"10","parent_id":"0","message":"Post","created":"2020-02-02T15:00:00Z","last_edited":"0001-01-01T00:00:00Z","allows_comments":true,"subspace":"desmos","creator":"cosmos1y54exmx84cqtasvjnskf9f63djuuj68p7hqf47","reactions":[{"owner":"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4","value":"like"},{"owner":"cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae","value":"like"}],"children":["98","100"]}`,
 		},
 	}
 
@@ -65,6 +82,61 @@ func TestPostQueryResponse_MarshalJSON(t *testing.T) {
 				test.expResponse,
 				string(jsonData),
 			)
+		})
+	}
+}
+
+func TestPostQueryResponse_String(t *testing.T) {
+	postOwner, _ := sdk.AccAddressFromBech32("cosmos1y54exmx84cqtasvjnskf9f63djuuj68p7hqf47")
+	liker, _ := sdk.AccAddressFromBech32("cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4")
+	otherLiker, _ := sdk.AccAddressFromBech32("cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae")
+
+	timeZone, _ := time.LoadLocation("UTC")
+	date := time.Date(2020, 2, 2, 15, 0, 0, 0, timeZone)
+	medias := types.PostMedias{
+		types.PostMedia{
+			URI:      "https://uri.com",
+			MimeType: "text/plain",
+		},
+	}
+
+	post := types.NewPost(
+		types.PostID(10),
+		types.PostID(0),
+		"Post",
+		true,
+		"desmos",
+		map[string]string{},
+		date,
+		postOwner,
+		medias,
+	)
+
+	likes := types.Reactions{
+		types.NewReaction("like", liker),
+		types.NewReaction("like", otherLiker),
+	}
+	children := types.PostIDs{types.PostID(98), types.PostID(100)}
+
+	PostResponse := types.NewPostResponse(post, likes, children)
+
+	tests := []struct {
+		name        string
+		response    types.PostQueryResponse
+		expResponse string
+	}{
+		{
+			name:        "Post query response string",
+			response:    PostResponse,
+			expResponse: "ID - [Reactions] [Children] \n10 - [[{\"owner\":\"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4\",\"value\":\"like\"} {\"owner\":\"cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae\",\"value\":\"like\"}]] [[98 100]]",
+		},
+	}
+
+	for _, test := range tests {
+		test := test
+		t.Run(test.name, func(t *testing.T) {
+			stringResponse := test.response.String()
+			assert.Equal(t, test.expResponse, stringResponse)
 		})
 	}
 }
